Move QualityOracle querying to 'core/quality' package
